+++ src/infrastructure/auth/auth.ts
@@ -4,7 +4,10 @@ import "server-only";
 import { betterAuth } from "better-auth";
 import { prismaAdapter } from "better-auth/adapters/prisma";
 
+import { getGoogleSocialProvider } from "@/infrastructure/auth/google-provider";
 import { prisma } from "@/infrastructure/prisma/client";
+
+const google = getGoogleSocialProvider();
 
 export const auth = betterAuth({
   database: prismaAdapter(prisma, {
@@ -14,6 +17,11 @@ export const auth = betterAuth({
     enabled: true,
     requireEmailVerification: false,
   },
+  socialProviders: google
+    ? {
+        google,
+      }
+    : {},
   secret: process.env.BETTER_AUTH_SECRET,
   baseURL: process.env.BETTER_AUTH_URL,
 });